requirements.txt
fastapi
uvicorn
groq
sentence-transformers
faiss-cpu
rank-bm25
sqlmodel
pypdf
python-docx
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
starlette==0.40.0

groq==0.9.0

sentence-transformers==3.0.1
faiss-cpu==1.8.0
rank-bm25==0.2.2
sqlmodel==0.0.22

pypdf==4.3.1
python-docx==1.1.2

reportlab==4.2.2
numpy==2.1.1
scikit-learn==1.5.2

requests==2.32.3
